requirements.txt
absl-py==2.2.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
Cython==3.0.12
daphne==4.1.2
Django==5.2
dlib==19.24.6
easydict==1.13
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
faiss-cpu==1.11.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
html5lib==1.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
insightface==0.7.3
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt_einsum==3.4.0
oscrypto==1.3.0
package_name==0.1
packaging==24.2
pillow==10.2.0
prettytable==3.16.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.5.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.0.0
reportlab==4.4.1
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
service-identity==24.2.0
simsimd==6.2.1
six==1.17.0
sounddevice==0.5.1
sqlparse==0.5.3
stringzilla==3.12.3
svglib==1.5.1
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==7.2
